Add force logout for session expired, Fix team dashboard schema

See merge request tanzeyan/sixth-sense-callio!154

+++ backend/schemas/metric.py
@@ -66,9 +66,9 @@ class AgentTable(BaseModel):
     leads: int
     followUps: int
     appointments: int
-    appointment_rate: int
+    appointment_rate: float
     bookings: int
-    booking_rate: int
+    booking_rate: float
     status: str
 
 class TeamStats(BaseModel):
